Bugfix: generator component key reference
_Why?_

When creating a component using the generator, the key to reference the
component was locked to be :text_block. This was a result of a careless
copy and paste. My bad... :(

_How?_

Updated the component template to manually write a key from the snake
cased name of the component.

_Side Effects?_

Hopefully, components created with the generator will be effective and
self documenting.